delete file phisically from files store and deletion of file from db + check is the user is onwer of the file!!!

+++ src/java/com/rest/controller/SystemController.java
@@ -127,6 +127,28 @@ public class SystemController {
         return dateFormat.format(date);
     }
 
+    private Long getUserId(String login){
+        
+        long userID = -1;
+        try {
+            DBConnector db = new DBConnector();
+            String sqlQuery = "select id from users where login='" + login + "'";
+
+            try (PreparedStatement statement = db.getConnection().prepareStatement(sqlQuery)) {
+                ResultSet rs = statement.executeQuery();
+                while (rs.next()) {
+                    userID = rs.getLong(1);
+                }
+            }
+            System.out.println("userid: " + userID);
+            db.closeConnection();
+
+        } catch (SQLException | ClassNotFoundException ex) {
+            Logger.getLogger(UsersController.class.getName()).log(Level.SEVERE, null, ex);
+        }
+        return userID;
+    }
+    
     public Long addFileInfoToDB(String fileName, Long fileSize, String tags, String path) {
 
         long id = 0;
@@ -164,18 +186,9 @@ public class SystemController {
 
     public void joinFileAndOwner(long fileID, String login) {
 
-        long userID = 0;
         try {
+            long userID = getUserId(login);
             DBConnector db = new DBConnector();
-            String sqlQuery = "select id from users where login='" + login + "'";
-
-            try (PreparedStatement statement = db.getConnection().prepareStatement(sqlQuery)) {
-                ResultSet rs = statement.executeQuery();
-                while (rs.next()) {
-                    userID = rs.getLong(1);
-                }
-            }
-            System.out.println("userid: " + userID + " fileid: " + fileID);
             String sqlQuery2 = "insert into files_users(userId, fileId, isOwner) values(?,?,?)";
             try (PreparedStatement statement = db.getConnection().prepareStatement(sqlQuery2)) {
                 statement.setLong(1, userID);
@@ -226,19 +239,10 @@ public class SystemController {
     }
 
     public void joinFolderAndOwner(Long folderID, String login) {
-        long userID = 0;
+
         try {
+            long userID = getUserId(login);
             DBConnector db = new DBConnector();
-            String sqlQuery = "select id from users where login='" + login + "'";
-
-            try (PreparedStatement statement = db.getConnection().prepareStatement(sqlQuery)) {
-                ResultSet rs = statement.executeQuery();
-                while (rs.next()) {
-                    userID = rs.getLong(1);
-                }
-            }
-
-            System.out.println("userid: " + userID + " folderid: " + folderID);
             String sqlQuery2 = "insert into folders_users(userId, folderId, isOwner) values(?,?,?)";
             try (PreparedStatement statement = db.getConnection().prepareStatement(sqlQuery2)) {
                 statement.setLong(1, userID);
@@ -247,7 +251,6 @@ public class SystemController {
                 statement.executeUpdate();
                 statement.close();
             }
-
             db.closeConnection();
 
         } catch (SQLException | ClassNotFoundException ex) {
@@ -255,38 +258,58 @@ public class SystemController {
         }
     }
 
-    public void deleteFileFromDB(String path, String fileName) {
-        //TODO: sprawdzić!!!! 
-        String sqlQuery = "";
+    public boolean deleteFileFromDB(String path, String fileName, String login) {
+
+        String sqlQuery;
+        String isOwner = "FALSE";
         Long fileID = Long.valueOf(-1);
+        Long userID;
+        
         try {
             DBConnector db = new DBConnector();
+            System.out.println("check, path:" + path + " filename:" + fileName + " login:" + login);
             sqlQuery = "select id from files where fileName='" + fileName + "' and directPath='" + path + "'";
-
             try (PreparedStatement statement = db.getConnection().prepareStatement(sqlQuery)) {
                 ResultSet rs = statement.executeQuery();
                 while (rs.next()) {
                     fileID = rs.getLong(1);
                 }
             }
-            System.out.println("DELETE: fileID " + fileID + "\npath: " + path + "\nfilename: " + fileName);
+            
+            //checks onwership:
+            userID = getUserId(login);
+            System.out.println("userid: " + userID + " fileID: " + fileID);
+            
+            sqlQuery = "select isOwner from files_users where userId='" + userID + "' and fileId='" + fileID + "'";
+            try (PreparedStatement statement = db.getConnection().prepareStatement(sqlQuery)) {
+                ResultSet rs = statement.executeQuery();
+                while (rs.next()) {
+                    isOwner = rs.getString(1);
+                }
+            }
+            
+            System.out.println("isOwner: " + isOwner);
+            if(isOwner.equals("FALSE")) {
+                return false;
+            }
+                       
+            //System.out.println("DELETE: fileID " + fileID + "\npath: " + path + "\nfilename: " + fileName);
             sqlQuery = "delete from files where fileName='" + fileName + "' and directPath='" + path + "'";
-
             try (PreparedStatement statement = db.getConnection().prepareStatement(sqlQuery)) {
                 statement.executeUpdate();
             }
 
             sqlQuery = "delete from files_users where fileId='" + fileID + "'";
-
             try (PreparedStatement statement = db.getConnection().prepareStatement(sqlQuery)) {
                 statement.executeUpdate();
             }
 
             db.closeConnection();
-
+            return true;         
         } catch (SQLException | ClassNotFoundException ex) {
             Logger.getLogger(UsersController.class.getName()).log(Level.SEVERE, null, ex);
         }
+        return false;
     }
 
     public boolean deleteFile(String path, String fileName) {
